Move Carousel component to shared
