+++ frontend/src/shared/components/Carousel/Carousel.less
