+++ frontend/src/shared/components/Carousel/index.tsx
